Update README.md
Avengers: Endgame is the direct sequel to Infinity War and the culmination of the MCU's Infinity Saga. After Thanos's snap erased half of all life, the remaining Avengers attempt to find a solution. Five years later, the discovery of time travel through the Quantum Realm allows the heroes to execute a "Time Heist" to collect the six Infinity Stones from the past in a final, desperate attempt to reverse the devastating damage and restore the universe.